[tf.data service] Fix issue with dynamic sharding of GroupByWindow.
Previously, we incorrectly propagated split providers into iterators created for the nested reduce_fn datasets in GroupByWindow. This could cause errors if a GroupByWindow uses a reduce_fn that produces a splittable dataset.

This CL also fixes a segmentation fault that previously happened in split_utils when CreateInputIteratorContexts received a context with too few split providers.

PiperOrigin-RevId: 431040630